Actions and reducers
Continuing work on achieving CRUD MVP. Update is the only step that still needs to be implemented. Some mapping of response data has changed. Will need to be updated from the hardcoded messages in the reducers, to displaying the response messages from BE. Log out, Delete Profile, Log in, Register are in the final phases of prototyping.