Update
Bug is still present. Unsure if this will occur in deploy. Appears to be caused by local document mutation that causes application to hotwire redux store to blank out, then attempts to make a GET from the auth route of home for a regUser and returns null on the GET request. does not cause total break of page, just will not allow for rendering of user data card. Propose temp backdoor to render conditional statement with reinforcing reducer/action response if continues to be an issue. 

Implemented the horizontal bar graph legend. Created as a custom react component now, can handle props passed to it. Will continue to work towards expected end goals for design and style of traits legend.